refactor: move logging config into the shared common package
Consolidate on the portal-assistant variant: request-id context/formatter,
the precise healthcheck access-log filter, idempotent filter install, and
the OpenAI debug-log toggle. setup_logging is parameterized so common
stays independent of per-agent settings; each agent's shim feeds its own
config (finops gains the openai_debug_logs field, default off).
